fix(storage): cast Postgres next_sort to BIGINT and fill empty 5xx bodies
Two reported errors:

1. \"mismatched types; Rust type i64 (as SQL type INT8) is not
   compatible with SQL type INT4\" on save_document.
   COALESCE(MAX(sort_order), -1) + 1 returned INT4 because
   sort_order is INTEGER; the Rust code expects i64. Cast each of
   the four next_sort selects to BIGINT explicitly.

2. /admin/* and /metering/summary returned 500s with empty bodies,
   making it impossible to diagnose from the browser. Many handlers
   in routes/{admin,dashboard,auth,account}.rs return
   StatusCode::INTERNAL_SERVER_ERROR.into_response() after logging
   the real cause; the response carried no body.

   Add a fill_empty_error_body middleware that buffers each 4xx/5xx
   response and, if the body is empty, fills it with a JSON
   {\"error\": <canonical reason>}. Responses that already have a
   body are passed through unchanged. dashboard.rs's metering
   handler additionally serializes the underlying error so the
   browser sees the specific cause.

+++ apps/backend/crates/kaya-storage/src/postgres.rs
@@ -639,7 +639,7 @@ async fn next_doc_sort_order_postgres(
 ) -> Result<i64, StorageError> {
     let row = match folder_id {
         Some(fid) => sqlx::query(
-            "SELECT COALESCE(MAX(sort_order), -1) + 1 AS next_sort \
+            "SELECT (COALESCE(MAX(sort_order), -1) + 1)::BIGINT AS next_sort \
              FROM documents WHERE deleted_at IS NULL AND folder_id = $1",
         )
         .bind(fid.to_string())
@@ -647,7 +647,7 @@ async fn next_doc_sort_order_postgres(
         .await
         .map_err(box_err)?,
         None => sqlx::query(
-            "SELECT COALESCE(MAX(sort_order), -1) + 1 AS next_sort \
+            "SELECT (COALESCE(MAX(sort_order), -1) + 1)::BIGINT AS next_sort \
              FROM documents WHERE deleted_at IS NULL AND folder_id IS NULL",
         )
         .fetch_one(pool)
@@ -723,7 +723,7 @@ async fn next_folder_sort_order_postgres(
 ) -> Result<i64, StorageError> {
     let row = match parent_id {
         Some(parent_id) => sqlx::query(
-            "SELECT COALESCE(MAX(sort_order), -1) + 1 AS next_sort
+            "SELECT (COALESCE(MAX(sort_order), -1) + 1)::BIGINT AS next_sort
              FROM folders
              WHERE parent_id = $1",
         )
@@ -732,7 +732,7 @@ async fn next_folder_sort_order_postgres(
         .await
         .map_err(box_err)?,
         None => sqlx::query(
-            "SELECT COALESCE(MAX(sort_order), -1) + 1 AS next_sort
+            "SELECT (COALESCE(MAX(sort_order), -1) + 1)::BIGINT AS next_sort
              FROM folders
              WHERE parent_id IS NULL",
         )
